Tertiary Commit
I remembered that Google Poly exists at the end of last class, but all of the models that I downloaded were gLTFs so now I am re-downloading them as OBJs which Unity can actually use. That is why there isn't much in there right now, but I do have a bunch of assets that I can add so it should be quick/easy to add a lot more in.